Allow using currency other than default.
Say we are using two charts of accounts, each with different currencies. Or, for some reason, we want to use a currency other than Money.default_currency.

This would allow one to use a different currency for the entire account, without having some balances returned in the default currency instead of the currency that all of the amounts are in.

Not sure if this is added complexity for the sake of supporting another currency, which shouldn't really be in the same set of accounts?

Perhaps the better solution, now that I think about it, is to implement the ability to define charts of accounts, each with their own base currency, then enforce the currency on any amoutns that are added to any of the accounts within it.